+++ src/mod.cpp
@@ -58,11 +58,13 @@ __declspec(dllexport) void __cdecl Init(const char* path, const HelperFunctions&
 
     archipelagoManager.SetServerConfiguration(serverIp, playerName, serverPassword);
 
+    const bool completeMultipleLevelMissions = settingsIni->getBool("GameSettings", "CompleteMultipleLevelMissions", true);
     const bool autoSkipCutscenes = settingsIni->getBool("GameSettings", "AutoSkipCutscenes", true);
     const bool skipCredits = settingsIni->getBool("GameSettings", "SkippeableCredits", true);
     const bool winButtonEnabled = settingsIni->getBool("GameSettings", "AutoWinButton", false);
 
     cheatsManager.SetCheatsConfiguration(autoSkipCutscenes, skipCredits, winButtonEnabled);
+    eventDetector.SetMultipleMissions(completeMultipleLevelMissions);
 
     char pathbuf[MAX_PATH];
     ReplacePNG_Common("HYOJI_EMBLEM0");