+++ CHANGELOG.md
@@ -5,9 +5,12 @@ All notable changes to this project will be documented in this file.
 ## [v0.2.3] - 2024-07-16
 
 ### Added
-- The mod will check for the requirements of other missions after finishing the level and give you the extra location checks.
+
+- The mod will check for the requirements of other missions after finishing the level and give you the extra location
+  checks.
 - Getting 3000 points on hedgehog hammer will give you both the check for Warrior's Feather and Long Hammer.
 - Added error message if the steam achievement mod is loaded.
+- The door between the hotel and the casino area will be open after getting both keys.
 
 ## [v0.2.2] - 2024-07-15